% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add two numbers}
\usage{
add(x, y)
}
\arguments{
\item{x}{The first number to add.}

\item{y}{The second number to add.}
}
\value{
The sum of two numbers
}
\description{
Add two numbers
}
\examples{
add(1, 2) # 3
add(4, 6) # 10
}
