% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umbrella.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{chisq.test}
\alias{aggregate}
\alias{cov}
\alias{example}
\alias{read.delim}
\alias{menu}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{stats}{\code{\link[stats]{aggregate}}, \code{\link[stats]{chisq.test}}, \code{\link[stats:cor]{cov}}}

  \item{utils}{\code{\link[utils]{example}}, \code{\link[utils]{menu}}, \code{\link[utils:read.table]{read.delim}}}
}}

