% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpkg3-package.R
\docType{package}
\name{testpkg3-package}
\alias{testpkg3}
\alias{testpkg3-package}
\title{testpkg3: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ramnathv/testpkg3}
  \item \url{https://ramnathv.github.io/testpkg3/}
  \item Report bugs at \url{https://github.com/ramnathv/testpkg3/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
